Move config directory into lib
