+++ packages/terra-functional-testing/lib/test-runner/index.js
@@ -1,5 +1,5 @@
 const TestRunner = require('./test-runner');
-const wdioConfig = require('../../config/wdio.conf');
+const wdioConfig = require('../config/wdio.conf');
 
 module.exports = {
   TestRunner,
