+++ packages/terra-functional-testing/lib/test-runner/test-runner.js
@@ -20,7 +20,7 @@ class TestRunner {
       return relativeConfigPath;
     }
 
-    return path.resolve(__dirname, '../../config/wdio.conf.js');
+    return path.resolve(__dirname, '../config/wdio.conf.js');
   }
 
   /**
