+++ packages/terra-functional-testing/tests/jest/config/wdio.conf.test.js
@@ -1,4 +1,4 @@
-const config = require('../../../config/wdio.conf.js');
+const config = require('../../../lib/config/wdio.conf.js');
 
 describe('WDIO Config', () => {
   it('should export the default WDIO configuration', () => {
